+++ Assets/lilToon/CHANGELOG.md
@@ -7,15 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.13] - 2022-xx-xx
 ### Added
 - `Receive Shadow` to 2nd / 3rd shadows
+- Added `Fully automatic setting` to shader setting (WIP)
 
 ### Changed
 - Improved alpha mask GUI
 - Changed the `Fix width by distance` of the outline to be adjustable steplessly
+- Adjusted the initial value of `Light Direction Override`
 
 ### Fixed
 - Fixed an issue where fog wasn't working well in URP
 - Fixed baking process when using other than PNG / JPG format for texture
 - Added `[MainTexture]` to the `_MainTex` property
+- Fixed an issue where `Light Direction Override` was affecting ShadeSH9 calculations
 
 ## [1.2.12] - 2022-03-31
 ### Fixed